+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added support for loading and previewing root motion translation for animations.
 * Added `merge-databases` command to xc3_shader for creating combined databases.
 * Added support to xc3_gltf for loading `monolib/shader` images.
+* Added support for rendering `.wiidcm` and `.idcm` collision files to xc3_viewer.
 
 ### Changed
 * Optimized shader database size.